requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17
amqp==5.0.6
asgiref==3.3.0
async-timeout==3.0.1
attrs==20.3.0
Automat==20.2.0
autopep8==1.5.4
beautifulsoup4==4.8.0
billiard==3.6.4.0
bs4==0.0.1
capmonster-python==1.3.0
celery==5.1.0
certifi==2020.4.5.1
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
Django==3.1.3
django-celery-results==2.0.1
django-six==1.0.4
djangorestframework==3.12.4
emoji==0.6.0
et-xmlfile==1.0.1
future==0.18.2
h2==3.2.0
hpack==3.0.0
httpie==2.4.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
instaloader==4.5.3
isodate==0.6.0
itemadapter==0.2.0
itemloaders==1.0.4
jdcal==1.4.1
jmespath==0.10.0
kombu==5.1.0
lxml==4.6.3
MechanicalSoup==1.0.0
msrest==0.6.21
multidict==5.1.0
numpy==1.19.3
oauthlib==3.1.1
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
openpyxl==3.0.5
pandas==1.1.4
parsel==1.6.0
pefile==2019.4.18
pika==1.2.0
Pillow==8.2.0
priority==1.3.0
prompt-toolkit==3.0.18
Protego==0.1.16
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.8.1
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
pyOpenSSL==20.0.1
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.8.1
python-weather==0.2.3
python3-anticaptcha==1.7.1
pytube==9.6.0
pytz==2020.4
pywin32==228
pywin32-ctypes==0.2.0
queuelib==1.5.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
Scrapy==2.5.0
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
typing-extensions==3.7.4.3
urllib3==1.25.10
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
win10toast==0.9
WMI==1.5.1
xmltodict==0.12.0
yarl==1.6.3
zope.interface==5.4.0
